day4accountpagefinished
finished 3rd section of page (favorites). Created display flexes to the container of all the containers inside the big container. Created a class for icons so they can be red and added a UL class so that they can be closer together. Made the images smaller using width, height, and object-fit: contain